refactor(components): efactor Logo, Navbar, Social, SettingPanel, SettingButton
Extract ALT_TEXT for Logo to i18n files. Convert Logo to Server Component when language is static.
Refactor Navbar, Social, SettingPanel and SettingButton for type safety and reusability.
