--- src/components/atom/media/index.ts
@@ -1,14 +1,0 @@
-/**
- * Archivo de barril (barrel file) para exportar componentes relacionados con medios visuales
- * desde el directorio actual.
- *
- * Facilita la importación centralizada y mejora la organización del código,
- * permitiendo importar componentes de medios desde un único punto.
- * 
- * Actualmente exporta:
- * - Logo: Componente para mostrar el logotipo de la aplicación.
- * - Avatar: Componente para mostrar el avatar o imagen de perfil de usuario.
- */
-
-export * from './Logo';
-export * from './Avatar';
