+++ src/components/templates/AuthLayout/AuthLayout.tsx
@@ -1,6 +1,6 @@
 import React, { ReactNode } from 'react';
 import Header from '@/components/organisms/Headers/Header';
-import { Sidebar } from '@/components/molecules/Navbar';
+import { Sidebar } from '@/components/molecules/navbar';
 import styles from './AuthLayout.module.scss';
 
 // Defining the structure of the links that can be passed as props
